Fix TOTP digit overflow and unread lockedMinutes on MFA rejection
codeAt's digit modulus was a uint32, which overflows once digits
reaches 10 (the max an otpauth:// URL may request per
parseTOTPSecret's own validation): 10^10 wraps in a uint32, producing
a wrong code for roughly a third of possible values. Widened to
uint64, with an independently-computed regression test.

mfaRejection.LockedMinutes was decoded from a -30139 response but
never read, so a controller reporting the account is already locked
still burned a second login attempt instead of stopping immediately
and naming the lockout.

+++ internal/omada/totp.go
@@ -120,9 +120,11 @@ func (t *totpConfig) codeAt(at time.Time) string {
 	sum := mac.Sum(nil)
 
 	offset := sum[len(sum)-1] & 0x0f
-	value := binary.BigEndian.Uint32(sum[offset:offset+4]) & 0x7fffffff
+	value := uint64(binary.BigEndian.Uint32(sum[offset:offset+4]) & 0x7fffffff)
 
-	mod := uint32(1)
+	// uint64, not uint32: digits can be up to 10 (see parseTOTPSecret), and
+	// 10^10 overflows a uint32.
+	mod := uint64(1)
 	for range t.digits {
 		mod *= 10
 	}
